feat: implement token bucket rate limiter with automatic recovery
Replaces adaptive rate limiter with token bucket algorithm:
- Proper fractional rate handling (0.1 req/s works correctly)
- Automatic recovery after cool-down period
- Cached rate limits persist to database across sessions
- Background recovery tasks for gradual rate increase
- Configurable min/max rates, ramp-up and back-off factors

New rate_limiter_v2.rs module with:
- TokenBucket struct with refill mechanics
- RecoveryConfig for cool-down and gradual recovery
- spawn_recovery_task() for background monitoring

Database migration uses DOUBLE PRECISION for f64 compatibility.

+++ migrations_postgres/20260119000009_add_rpc_rate_limits_table.sql
@@ -13,17 +13,17 @@ CREATE TABLE IF NOT EXISTS rpc_rate_limits (
     rpc_url TEXT NOT NULL UNIQUE,
 
     -- Rate limit information
-    discovered_rate_limit REAL NOT NULL,
+    discovered_rate_limit DOUBLE PRECISION NOT NULL,
     rate_limit_strategy TEXT NOT NULL, -- 'manual', 'adaptive', 'aggressive'
 
     -- Statistics
     total_requests BIGINT NOT NULL DEFAULT 0,
     rate_limit_errors BIGINT NOT NULL DEFAULT 0,
-    success_rate REAL NOT NULL DEFAULT 1.0,
+    success_rate DOUBLE PRECISION NOT NULL DEFAULT 1.0,
 
     -- Measurements
-    measured_rps REAL,
-    avg_response_time_ms REAL,
+    measured_rps DOUBLE PRECISION,
+    avg_response_time_ms DOUBLE PRECISION,
 
     -- Timestamps
     first_seen_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
